Removed __init__.py from repo root
The __init__.py was moved from the previously existing sub-directory
`caejobdiary` into the the repo root, when the sub-directory was
removed.

This caused issues with imports when the project directory is called
`caejobdiary` -- which happened to be the case on the CI.
On my machine the repo is called `CAEJobDiary`.

For some reason on the CI when running the tests it was looking for a
`caejobdiary.diary` module rather than just a `diary` module. Not sure
why because the import is done like `from diary.models import ...`.

With the `__init__.py` removed from the repo root, the tests run without
importing issues -- no matter what name the repository has.