Created simple HTML form to send tags as array
This is just the very basic implementation and does not allow
to make use of the form validation that other Django form fields
provide. Also, this would require a bunch of work in the view.

It probably makes more sense to create a custom form for how I
envision the tags should be displayed and edited in the job detail.