Commented and simplified detail view flow
Added some comments that should make it easier to understand
what is happening in the future.

Also change the flow a little to actually enable the display of
form error on the page. So far it does not quite seem possible
to have any form errors, because the `select` elements only show
valid answers and the `textarea` elements have no real contraints.

Hm? what is happening when I try too many characters?