Storing joblist querystring in session
Before, the querystring from the joblist was attached to all
outgoing links (e.g. the detail views). This meant the details views
had a URL with query parameters in it, that have no meaningful effect
for the detail view. The query parameters were only used in the detail
view to create links back to the filtered list view.

To remove the confusing use of the queryparameters in the detail view,
the functionality has been switched. The joblist view stores its
querystring in the session. This value is then retrieved in the detail
view and passed to the detail template to generate the return links.

Not more queryparameters are present in the URL for the detail page.